Resolve reference to link_to within controller.
There seems to be some indications that using "link_to" within a
controller within a controller is bad practice. Most indications seem to
be that you could hand-roll the HTML yourself (a thousand times worse,
in my opinion), render a partial (which I don't know how to do), or just
leave the link_to for views.

However, outside of extending this to include a new view which is then
rendered and passed in to the flash notice, I think this is the quickest
and sanest option.

If it actually works!

Ref: http://blog.lrdesign.com/2009/05/using-link_to-or-other-helper-methods-in-a-controller/